+++ src/pages/StatsPage.tsx
@@ -3,19 +3,9 @@ import { useSearchParams } from 'react-router-dom';
 import useLoadCost from '../hooks/useLoadCost';
 import { Chart as ChartJS, registerables } from 'chart.js';
 import { Line } from 'react-chartjs-2';
+import {options, containerStyles } from '../components/statis/statCss'
 
 ChartJS.register(...registerables);
-
-type ChartDataType = {
-  labels: string[];
-  datasets: {
-    label: string;
-    data: number[];
-    backgroundColor: string;
-    borderColor: string;
-    borderWidth: number;
-  }[];
-};
 
 const StatsPage = () => {
   const [searchParams] = useSearchParams();
@@ -34,72 +24,6 @@ const StatsPage = () => {
   });
   const [period, setPeriod] = useState<'week' | 'month'>('week');
 
-  const containerStyles: React.CSSProperties = {
-    height: '600px',
-    width: '1800px',
-    marginTop: '80px',
-    marginLeft: '40px',
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    background:
-      'linear-gradient(180deg, rgba(255, 165, 0, 0.8), rgba(128, 0, 128, 0.8))',
-    borderRadius: '15px',
-    boxShadow: '0px 4px 10px rgba(0, 0, 0, 0.25)',
-    padding: '20px',
-  };
-
-  const options = {
-    responsive: true,
-    maintainAspectRatio: false,
-    scales: {
-      y: {
-        beginAtZero: true,
-        ticks: {
-          color: '#fff',
-        },
-        grid: {
-          color: 'rgba(255, 255, 255, 0.1)',
-        },
-      },
-      x: {
-        offset: true,
-        ticks: {
-          color: 'gold',
-          font: {
-            size: 14,
-            weight: 'bold' as const,
-          },
-          textStrokeColor: 'black',
-          textStrokeWidth: 3,
-        },
-        afterFit: (scaleInstance: { height: number }) => {
-          scaleInstance.height = scaleInstance.height * 2.2;
-        },
-      },
-    },
-    plugins: {
-      legend: {
-        position: 'top' as const,
-        labels: {
-          color: '#fff',
-        },
-      },
-    },
-    elements: {
-      point: {
-        radius: 5,
-      },
-      line: {
-        tension: 0.4,
-      },
-    },
-  };
-
-  type RecordType = 'expenses' | 'income';
-  type CategoryRecord = ExpenseRecord | IncomeRecord;
-
   const processChartData = useCallback(
     (
       records: CategoryRecord[],
@@ -107,7 +31,7 @@ const StatsPage = () => {
       categoryId: string,
     ): ChartDataType => {
       const startDate = new Date();
-      startDate.setHours(0, 0, 0, 0); // Установка времени на начало дня
+      startDate.setHours(0, 0, 0, 0); 
       const endDate = new Date();
       if (period === 'month') {
         endDate.setMonth(startDate.getMonth() + 1);
